git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/trunk@1006 8c327b6c-a73e-8745-8738-111f1e176ae1

+++ src/AGC/AGCEventDef.cpp
@@ -300,7 +300,7 @@ HRESULT CAGCEventDef::ExpandFmtString(BSTR bstrFmt, IAGCEvent* pEvent,
   LARGE_INTEGER li = {0};
   ULARGE_INTEGER uli;
   RETURN_FAILED(spStm->Seek(li, STREAM_SEEK_CUR, &uli));
-  UINT cchStream = uli.QuadPart / sizeof(OLECHAR);
+  UINT cchStream = (UINT)uli.QuadPart / sizeof(OLECHAR);
 
   // Get the HGLOBAL underlying the stream
   HGLOBAL hGlobal = NULL;
